+++ build.gradle.kts
@@ -1,49 +1,17 @@
 plugins {
 	java
-	id("org.springframework.boot") version "3.2.1"
-	id("io.spring.dependency-management") version "1.1.4"
+  alias(libs.plugins.spring.boot) apply false
+  alias(libs.plugins.spring.dependencies) apply false
 }
 
-group = "com.muhrifqii"
-version = "0.0.1-SNAPSHOT"
+allprojects {
+  repositories {
+    mavenCentral()
+    maven(url = "https://repo.spring.io/milestone")
+    maven(url = "https://repo.spring.io/snapshot")
+  }
 
-java {
-	sourceCompatibility = JavaVersion.VERSION_17
-}
-
-configurations {
-	compileOnly {
-		extendsFrom(configurations.annotationProcessor.get())
-	}
-}
-
-repositories {
-	mavenCentral()
-	maven(url = "https://repo.spring.io/milestone")
-	maven(url = "https://repo.spring.io/snapshot")
-}
-
-dependencies {
-	implementation("org.springframework.boot:spring-boot-starter-data-r2dbc")
-	implementation("org.springframework.boot:spring-boot-starter-webflux")
-	implementation("org.flywaydb:flyway-core")
-	implementation("org.springframework:spring-jdbc")
-
-	implementation("org.springframework.ai:spring-ai-ollama-spring-boot-starter:0.8.0-SNAPSHOT")
-
-	compileOnly("org.projectlombok:lombok")
-
-	runtimeOnly("org.postgresql:postgresql")
-	runtimeOnly("org.postgresql:r2dbc-postgresql")
-
-	annotationProcessor("org.projectlombok:lombok")
-
-	testImplementation("org.springframework.boot:spring-boot-starter-test")
-	testImplementation("org.springframework.boot:spring-boot-testcontainers")
-	testImplementation("org.testcontainers:junit-jupiter")
-	testImplementation("io.projectreactor:reactor-test")
-}
-
-tasks.withType<Test> {
-	useJUnitPlatform()
+  tasks.withType<Test> {
+    useJUnitPlatform()
+  }
 }
