+++ buildSrc/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+  `kotlin-dsl`
+}
+
+java {
+	sourceCompatibility = JavaVersion.VERSION_22
+}
