+++ gradle.properties
@@ -1,0 +1,3 @@
+org.gradle.caching=true
+org.gradle.parallel=true
+org.gradle.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
