+++ gradle/libs.versions.toml
@@ -1,0 +1,11 @@
+[versions]
+spring-boot = "3.3.2"
+spring-ai = "1.0.0-SNAPSHOT"
+
+[libraries]
+spring-ai-bom = { group = "org.springframework.ai", name = "spring-ai-bom", version.ref = "spring-ai" }
+spring-ai-ollama = { group = "org.springframework.ai", name = "spring-ai-ollama-spring-boot-starter" }
+
+[plugins]
+spring-boot = { id = "org.springframework.boot", version.ref = "spring-boot" }
+spring-dependencies = { id = "io.spring.dependency-management", version = "1.1.4" }
