+++ llm/build.gradle.kts
@@ -1,0 +1,46 @@
+plugins {
+	java
+  alias(libs.plugins.spring.boot)
+  alias(libs.plugins.spring.dependencies)
+}
+
+group = "com.muhrifqii"
+version = "0.0.1-SNAPSHOT"
+
+java {
+	sourceCompatibility = JavaVersion.VERSION_22
+}
+
+configurations {
+	compileOnly {
+		extendsFrom(configurations.annotationProcessor.get())
+	}
+}
+
+dependencies {
+	implementation("org.springframework.boot:spring-boot-starter-data-r2dbc")
+	implementation("org.springframework.boot:spring-boot-starter-webflux")
+	implementation("org.flywaydb:flyway-core")
+	implementation("org.springframework:spring-jdbc")
+
+  implementation(platform(libs.spring.ai.bom))
+	implementation(libs.spring.ai.ollama)
+
+	compileOnly("org.projectlombok:lombok")
+
+	runtimeOnly("org.postgresql:postgresql")
+	runtimeOnly("org.postgresql:r2dbc-postgresql")
+
+  developmentOnly("org.springframework.boot:spring-boot-devtools")
+
+	annotationProcessor("org.projectlombok:lombok")
+
+	testImplementation("org.springframework.boot:spring-boot-starter-test")
+	testImplementation("org.springframework.boot:spring-boot-testcontainers")
+	testImplementation("org.testcontainers:junit-jupiter")
+	testImplementation("io.projectreactor:reactor-test")
+}
+
+tasks.test {
+  useJUnitPlatform()
+}
