+++ llm/src/main/java/com/muhrifqii/llm/SpringbedApplication.java
@@ -1,0 +1,15 @@
+package com.muhrifqii.llm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.r2dbc.config.EnableR2dbcAuditing;
+
+@SpringBootApplication
+@EnableR2dbcAuditing
+public class SpringbedApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringbedApplication.class, args);
+	}
+
+}
