+++ llm/src/main/java/com/muhrifqii/llm/dto/ChatMessageRequest.java
@@ -1,0 +1,4 @@
+package com.muhrifqii.llm.dto;
+
+public record ChatMessageRequest(String message) {
+}
