+++ llm/src/main/java/com/muhrifqii/llm/dto/CreateBedRequest.java
@@ -1,0 +1,7 @@
+package com.muhrifqii.llm.dto;
+
+public record CreateBedRequest(
+                String name,
+                String description,
+                String category) {
+}
