+++ llm/src/main/java/com/muhrifqii/llm/dto/ErrorHttpResponse.java
@@ -1,0 +1,7 @@
+package com.muhrifqii.llm.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ErrorHttpResponse(int status, int code, String message, String timestamp) {
+}
