+++ llm/src/main/java/com/muhrifqii/llm/exceptions/InvalidRequestException.java
@@ -1,0 +1,19 @@
+package com.muhrifqii.llm.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class InvalidRequestException extends RuntimeException {
+
+    private final int code;
+
+    public InvalidRequestException(Throwable cause, int code) {
+        super(cause);
+        this.code = code;
+    }
+
+    public InvalidRequestException(String message, int code) {
+        super(message);
+        this.code = code;
+    }
+}
