+++ llm/src/main/java/com/muhrifqii/llm/repositories/BedRepository.java
@@ -1,0 +1,21 @@
+package com.muhrifqii.llm.repositories;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.repository.reactive.ReactiveCrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.muhrifqii.llm.domains.Bed;
+
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+@Repository
+public interface BedRepository extends ReactiveCrudRepository<Bed, Long> {
+    Flux<Bed> findAllBy(Pageable pageable);
+
+    Flux<Bed> findAllByCategory(Bed.Category category, Pageable pageable);
+
+    Mono<Bed> findByIdentifier(String identifier);
+
+    int countByCategory(Bed.Category category);
+}
