+++ llm/src/main/java/com/muhrifqii/llm/utils/DateUtils.java
@@ -1,0 +1,15 @@
+package com.muhrifqii.llm.utils;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class DateUtils {
+    public static String now() {
+        return LocalDateTime.now()
+                .format(DateTimeFormatter.ISO_DATE_TIME);
+    }
+}
