+++ llm/src/main/java/com/muhrifqii/llm/utils/StringUtils.java
@@ -1,0 +1,18 @@
+package com.muhrifqii.llm.utils;
+
+import jakarta.annotation.Nonnull;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import java.util.Arrays;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class StringUtils {
+    public static <T extends Enum<T>> T enumValueOf(String name, @Nonnull Class<T> type) {
+
+        return Arrays.stream(type.getEnumConstants())
+                .filter(e -> e.name().equalsIgnoreCase(name))
+                .findAny()
+                .orElseThrow(() -> new IllegalArgumentException("No enum constant"));
+    }
+}
