+++ llm/src/main/resources/migration/V1__initial_setup.sql
