+++ llm/src/test/java/com/muhrifqii/llm/SpringbedApplicationTests.java
@@ -1,0 +1,13 @@
+package com.muhrifqii.llm;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringbedApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
