+++ settings.gradle.kts
@@ -1,1 +1,9 @@
+pluginManagement {
+  repositories {
+    gradlePluginPortal()
+    mavenCentral()
+  }
+}
+
 rootProject.name = "SlakingSpringBedAI"
+include("llm")
