--- src/main/java/com/muhrifqii/springbed/dto/ChatMessageRequest.java
@@ -1,4 +1,0 @@
-package com.muhrifqii.springbed.dto;
-
-public record ChatMessageRequest(String message) {
-}
