--- src/main/java/com/muhrifqii/springbed/dto/CreateBedRequest.java
@@ -1,8 +1,0 @@
-package com.muhrifqii.springbed.dto;
-
-public record CreateBedRequest(
-        String name,
-        String description,
-        String category
-) {
-}
