--- src/main/java/com/muhrifqii/springbed/dto/ErrorHttpResponse.java
@@ -1,7 +1,0 @@
-package com.muhrifqii.springbed.dto;
-
-import lombok.Builder;
-
-@Builder
-public record ErrorHttpResponse(int status, int code, String message, String timestamp) {
-}
