--- src/main/java/com/muhrifqii/springbed/utils/DateUtils.java
@@ -1,15 +1,0 @@
-package com.muhrifqii.springbed.utils;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class DateUtils {
-    public static String now() {
-        return LocalDateTime.now()
-                .format(DateTimeFormatter.ISO_DATE_TIME);
-    }
-}
