--- src/test/java/com/muhrifqii/springbed/SpringbedApplicationTests.java
@@ -1,13 +1,0 @@
-package com.muhrifqii.springbed;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringbedApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
